Fix warnings

+++ GUIApp/include/AMesh.hpp
@@ -14,7 +14,7 @@ public:
 	void init();
 	void draw();
 
-	virtual size_t getIndicesCount() = 0;
+	virtual int getIndicesCount() = 0;
 
 protected:
 	virtual void generateMesh(std::vector<glm::vec3>& vertices, std::vector<unsigned short>& indices) = 0;
